+++ src/mongo_client.ts
@@ -8,65 +8,7 @@ import { maybePromise, MongoDBNamespace } from './utils';
 import { deprecate } from 'util';
 import { connect, validOptions } from './operations/connect';
 import { PromiseProvider } from './promise_provider';
-import type { Callback, Document, BSONSerializeOptions } from './types';
-
-export const enum AutoEncryptionLoggerLevels {
-  FatalError = 0,
-  Error = 1,
-  Warning = 2,
-  Info = 3,
-  Trace = 4
-}
-
-export interface AutoEncryptionOptions {
-  /** A `MongoClient` used to fetch keys from a key vault */
-  keyVaultClient?: MongoClient;
-  /** The namespace where keys are stored in the key vault */
-  keyVaultNamespace?: string;
-  /** Configuration options that are used by specific KMS providers during key generation, encryption, and decryption. */
-  kmsProviders?: {
-    /** Configuration options for using 'aws' as your KMS provider */
-    aws?: {
-      /** The access key used for the AWS KMS provider */
-      accessKeyId?: string;
-      /** The secret access key used for the AWS KMS provider */
-      secretAccessKey?: string;
-    };
-    /** Configuration options for using 'local' as your KMS provider */
-    local?: {
-      /** The master key used to encrypt/decrypt data keys. A 96-byte long Buffer. */
-      key?: Buffer;
-    };
-  };
-  /**
-   * A map of namespaces to a local JSON schema for encryption
-   *
-   * **NOTE**: Supplying options.schemaMap provides more security than relying on JSON Schemas obtained from the server.
-   * It protects against a malicious server advertising a false JSON Schema, which could trick the client into sending decrypted data that should be encrypted.
-   * Schemas supplied in the schemaMap only apply to configuring automatic encryption for client side encryption.
-   * Other validation rules in the JSON schema will not be enforced by the driver and will result in an error.
-   */
-  schemaMap?: Document;
-  /** Allows the user to bypass auto encryption, maintaining implicit decryption */
-  bypassAutoEncryption?: boolean;
-  options?: {
-    /** An optional hook to catch logging messages from the underlying encryption engine */
-    logger?: (level: AutoEncryptionLoggerLevels, message: string) => void;
-  };
-  extraOptions?: {
-    /**
-     * A local process the driver communicates with to determine how to encrypt values in a command.
-     * Defaults to "mongodb://%2Fvar%2Fmongocryptd.sock" if domain sockets are available or "mongodb://localhost:27020" otherwise
-     */
-    mongocryptdURI?: string;
-    /** If true, autoEncryption will not attempt to spawn a mongocryptd before connecting  */
-    mongocryptdBypassSpawn?: boolean;
-    /** The path to the mongocryptd executable on the system */
-    mongocryptdSpawnPath?: string;
-    /** Command line arguments to use when auto-spawning a mongocryptd */
-    mongocryptdSpawnArgs?: string[];
-  };
-}
+import type { Callback, BSONSerializeOptions, AutoEncryptionOptions } from './types';
 
 export enum ReadPreferenceMode {
   primary = 'primary',
